refactor: Util 패키지 생성에 따른 HttpRequestUtil 이동
